[MTS/LT][93579030] Fix data extension row delete
The ExactTarget API specifies the use of "Key" and "Keys" elements
instead of "Property" and "Properties" when deleting a data extension
row. We now pass the action into the method that constructs the relevant
part of the SOAP request. That method decides what elements to use,
depending on what action is being performed. In the future, it may be
good to create some specialized objects to handle variations in
behavior.

`#add_customer_key` now assumes that the properties attribute is a hash.
When performing post, patch, or delete requests, the official
documentation uses a hash for the properties attribute. The old
implementation of the method assumed that the properties attribute was
an array.

Three tests fail. However, I am assuming the ExactTarget API changed at
one point. I am not sure what the desired output of the tested methods
is supposed to be now.